requirements-dev.txt
Red-DiscordBot
python-dotenv
twine
pytest
pytest-cov